chore(amp): 更新 Rockchip RT-Thread SDK
将 Rockchip RT-Thread 从 3.1.3 同步到 4.1.1。\n\n保留 RK3568 AMP 的 UART7、ISR_COUNT、GIC route、RPMsg-Lite 与 serial 生命周期兼容。

+++ components/amp/rockchip/rt-thread/bsp/rockchip/common/drivers/drv_clock.h
@@ -40,7 +40,7 @@ struct clk_gate
 struct clk_init
 {
     const char *name;
-    uint32_t clk_id;
+    uint64_t clk_id;
     uint32_t init_rate;
 };
